% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.stapDP.R
\name{as.array.stapDP}
\alias{as.array.stapDP}
\title{Retrieve DP parameter samples in Array Form}
\usage{
\method{as.array}{stapDP}(x, ...)
}
\description{
Retrieve DP parameter samples in Array Form
}
