% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.stapDP.R
\name{as.matrix.stapDP}
\alias{as.matrix.stapDP}
\title{Retrieve Parameter Samples in Matrix Form}
\usage{
\method{as.matrix}{stapDP}(x, ...)
}
\arguments{
\item{x}{a stapDP object}

\item{...}{ignored}
}
\description{
Returns the parameter samples from the model associated with the 
fixed effects, concentration parameter, residual variance, DP probabilities
and smooth scales as  a S X P matrix where S is the number of samples and P is the total number of parameters
}
\seealso{
as.array.stapDP for an array like container for the DP smooth components
}
