% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-summary.R
\name{diagnostics}
\alias{diagnostics}
\alias{diagnostics.stapDP}
\title{Diagnostics}
\usage{
diagnostics(x)

\method{diagnostics}{stapDP}(x)
}
\arguments{
\item{x}{a stapDP object}
}
\description{
Reports the Geweke diagnostic statistic for the concentration parameter alpha,
residual standard deviation sigma and cluster specific probabilities \eqn{\pi_k}.
}
\section{Methods (by class)}{
\itemize{
\item \code{stapDP}: 
}}

