% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_staplm.R
\name{fdp_staplm.fit}
\alias{fdp_staplm.fit}
\title{Functional Dirichlet Process Spatial Temporal Aggregated Predictor in a Linear Model}
\usage{
fdp_staplm.fit(y, Z, X, alpha_a, alpha_b, K, penalize, tau_df, w = rep(1,
  length(y)), iter_max, burn_in, thin, seed = NULL)
}
\arguments{
\item{y}{vector of outcomes}

\item{Z}{design matrix}

\item{X}{stap design matrix}

\item{alpha_a}{alpha gamma prior hyperparameter}

\item{alpha_b}{alpha gamma prior hyperparameter}

\item{K}{truncation number for DP mixture components}

\item{w}{weights for weighted regression - default is 1}

\item{iter_max}{maximum number of iterations}

\item{burn_in}{number of burn in iterations}

\item{thin}{number by which to thin samples}

\item{seed}{rng initializer}

\item{nu_0}{inverse-chi-square degree of freedom for tau base measure}
}
\description{
Functional Dirichlet Process Spatial Temporal Aggregated Predictor in a Linear Model
}
