% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_staplmer.R
\name{fdp_staplmer.fit}
\alias{fdp_staplmer.fit}
\title{Functional Dirichlet Process Spatial Temporal Aggregated Predictor Linear Mixed Effects Regression Model Fit}
\usage{
fdp_staplmer.fit(
  y,
  Z,
  X,
  W,
  subj_mat,
  subj_n,
  weights = rep(1, length(y)),
  alpha_a = 1,
  alpha_b = 1,
  sigma_a = 1,
  sigma_b = 1,
  tau_a = 1,
  tau_b = 1,
  K = 5L,
  threshold = 0L,
  iter_max,
  burn_in,
  thin = 1L,
  fix_alpha = FALSE,
  bw = FALSE,
  seed = NULL,
  chain = 1L,
  logging = FALSE,
  summarize_yhat = FALSE
)
}
\arguments{
\item{y}{vector of outcomes}

\item{Z}{design matrix}

\item{X}{stap design matrix}

\item{W}{group terms design matrix from \code{\link[lme4]{glFormula}}}

\item{subj_mat}{matrix indexing subject-measurement locations in (Z,X,W)}

\item{subj_n}{vector of number of subject measurements}

\item{weights}{weights for weighted regression - default is vector of ones}

\item{alpha_a}{alpha gamma prior hyperparameter}

\item{alpha_b}{alpha gamma prior hyperparameter}

\item{sigma_a}{precision gamma prior hyperparameter}

\item{sigma_b}{precision gamma prior hyperparameter}

\item{tau_a}{penalty parameters gamma prior hyperparameter}

\item{tau_b}{penalty parameters gamma prior hyperparameter}

\item{K}{truncation number for DP mixture components}

\item{threshold}{number of members per cluster at which cluster is included in regression}

\item{iter_max}{maximum number of iterations}

\item{burn_in}{number of iterations to burn-in}

\item{thin}{number by which to thin samples}

\item{fix_alpha}{boolean value}

\item{bw}{boolean value indicating whether or not subject decomposition is used}

\item{seed}{random number generator seed will be set to default value if not by user}

\item{chain}{chain label}

\item{logging}{boolean parameter indicating whether or not a single iteration should be run with print messages indicating successful completion of the Sampler's sub modules}

\item{summarize_yhat}{boolean value indicating whether a single mean vector of yhat values should be returned instead of a N X num samples matrix. Useful in situations where N is large.}
}
\description{
Functional Dirichlet Process Spatial Temporal Aggregated Predictor Linear Mixed Effects Regression Model Fit
}
