% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapDP-methods.R
\name{formula.stapDP}
\alias{formula.stapDP}
\title{Formula method for stapDP objects}
\usage{
\method{formula}{stapDP}(x, ...)
}
\arguments{
\item{...}{ignored currently}
}
\description{
Formula method for stapDP objects
}
\keyword{internal}
