% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapDPspec.R
\name{get_stapDPspec}
\alias{get_stapDPspec}
\title{Retrieve STAP-DP model specification from STAP model formula}
\usage{
get_stapDPspec(f, K, benvo)
}
\arguments{
\item{f}{formula from \code{\link{fdp_staplm}}, \code{\link{fdp_staplmer}}}

\item{K}{DP truncation integer}

\item{benvo}{Built Environment object - \code{\link[rbenvo]{Benvo}} - containing data for model}
}
\value{
\code{\link{stapDPspec}} object
}
\description{
Get stapDPspec object which details the various components of the stapDP model specification
}
