% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mode.R
\name{green_loss.stapDP}
\alias{green_loss.stapDP}
\title{Estimates Posterior Mode Loss}
\usage{
green_loss.stapDP(object, truth = NULL, tau = 0.5, a = 1, b = 1)
}
\arguments{
\item{object}{stapDP object}

\item{truth}{true adjacency matrix}

\item{tau}{loss function tuning parameter}

\item{a}{misclassification penalty parameter}

\item{b}{misclassification penalty parameter - see Green and Lau for details}
}
\value{
list of three objects (1)loss:  the loss corresponding to each clustering 
configuration, (2) best_loss_ix:  The index corresponding to the "best" partition,
and (3) mode: the best partition itself
}
\description{
Estimates Posterior Mode Loss
}
