% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{longitudinal_clusters}
\alias{longitudinal_clusters}
\title{Longitudinal Clustering benvo for use in \pkg{rstapDP} vignettes.}
\format{
A \code{\link[rbenvo]{benvo}} consisting of a subject and subject-distance dataframes
}
\usage{
longitudinal_clusters
}
\description{
Longitudinal Clustering benvo for use in \pkg{rstapDP} vignettes.
}
\details{
The outcome in the subject dataframe is simulated using a cluster specific effect as follows:
\deqn{E[Y_{ij}|b_{i1},b_{i2}] = 33  - 2.2 \text{sex}_i + .1*(year_{ij}) +  f_k(\text{FFR exposure}) +b_{i1} + b_{i2}(year_{ij})  } Where \eqn{k= 1,2,3} with varying probabilities.
}
\seealso{
the generating \href{https://github.com/apeterson91/rstapDP/tree/master/data-raw}{code} on Github.
}
\keyword{datasets}
