% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.stapDP}
\alias{plot.stapDP}
\title{Cluster-Specific Spatial Temporal Effects}
\usage{
\method{plot}{stapDP}(x, p = 0.95, switch = "color",
  prob_filter = 0.1, ...)
}
\arguments{
\item{x}{stapDP object}

\item{p}{probability contained in credible interval}

\item{switch}{one of "color" or "facet" for different plotting options}

\item{prob_filter}{all mixture components with median probability < prob_filter are excluded from the plot}

\item{...}{ignored}
}
\value{
plot with cluster effect across space
}
\description{
Cluster-Specific Spatial Temporal Effects
}
\keyword{internal}
