% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cluster_effects.stapDP}
\alias{plot_cluster_effects.stapDP}
\title{plot cluster effects plots}
\usage{
\method{plot_cluster_effects}{stapDP}(x, p = 0.95, switch = "color",
  prob_filter = 0.1)
}
\arguments{
\item{x}{stapDP object}
}
\description{
plot cluster effects plots
}
