% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pairs}
\alias{plot_pairs}
\alias{plot_pairs.stapDP}
\title{plots pairwise probability clustering plot}
\usage{
plot_pairs(x, sort = FALSE, sample = 0)

\method{plot_pairs}{stapDP}(x, sort = FALSE, sample = 0)
}
\arguments{
\item{x}{stapDP object}

\item{sort}{boolean asking whether sorting algorithm should be used to sort
pairwise probablity}

\item{sample}{number of subjects to sample in order to speed computation}
}
\value{
ggplot plot object
}
\description{
plots pairwise probability clustering plot
}
\section{Methods (by class)}{
\itemize{
\item \code{stapDP}: plot_pairwise probability matrix
}}

\references{
Rodriguez A., Dunson D, and Gelfand A. \emph{The Nested Dirichlet Process}
Journal of the American Statistical Association 103 (1131-1154).
}
\seealso{
the supplementary section of the reference for the sorting algorithm.
}
