% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pairs.stapDP}
\alias{plot_pairs.stapDP}
\title{plots pairwise probability clustering plot}
\usage{
\method{plot_pairs}{stapDP}(x, sort = FALSE)
}
\arguments{
\item{x}{stapDP object}

\item{sort}{boolean asking whether sorting algorithm should be used to sort}
}
\value{
ggplot plot object
}
\description{
plots pairwise probability clustering plot
}
