% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotdf}
\alias{plotdf}
\title{Cluster-Specific Spatial Temporal Effects  Plot Dataframe}
\usage{
plotdf(x, p = 0.95, prob_filter = 0.1)
}
\arguments{
\item{x}{stapDP object}

\item{p}{probability contained in credible interval}

\item{prob_filter}{all mixture components with median probability < prob_filter are excluded from the plot}
}
\value{
plot with cluster effect across space
}
\description{
Returns the plotting dataframe used for plotting the clustered spatial exposure effects
}
