% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-summary.R
\name{print.stapDP}
\alias{print.stapDP}
\title{Print and Summary Method}
\usage{
\method{print}{stapDP}(x, digits = 1)
}
\arguments{
\item{x}{stapDP object}

\item{digits}{number of digits to round off to}
}
\description{
Print and Summary Method
}
