% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mode.R, R/plots.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assign_mode}
\alias{green_loss}
\alias{plot_pairs}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bendr}{\code{\link[bendr]{assign_mode}}, \code{\link[bendr]{green_loss}}, \code{\link[bendr]{plot_pairs}}}
}}

