% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstapDP-package.R
\name{rsstap-package}
\alias{rsstap-package}
\alias{rsstap}
\title{The `rsstap package.`}
\description{
rsstap is a package that implements spatial-temporal aggregated predictor functions using splines in R. This allows for the modeling of features impact on measured subjects that can be related either through space or time.
}
