\name{rstapDP-package}
\alias{rstapDP-package}
\alias{rstapDP}
\docType{package}
\title{
\packageTitle{rstapDP}
}
\description{
\packageDescription{rstapDP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rstapDP}
\packageIndices{rstapDP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rstapDP}

Maintainer: \packageMaintainer{rstapDP}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
