% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{s_stap}
\alias{s_stap}
\title{STAP Smooth Term}
\usage{
s_stap(term, k)
}
\arguments{
\item{term}{single covariate name to be clustered}

\item{k}{basis dimension see \code{\link[mgcv]{s}}.}
}
\description{
STAP Smooth Term
}
