% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapDP-methods.R
\name{stapDP-methods}
\alias{stapDP-methods}
\alias{VarCorr}
\alias{ngrps}
\alias{sigma}
\alias{nsamples}
\alias{coef.stapDP}
\alias{confint.stapDP}
\alias{fitted.stapDP}
\alias{nobs.stapDP}
\alias{VarCorr.stapDP}
\alias{ngrps.stapDP}
\alias{nsamples.stapDP}
\title{Methods for stapDP objects}
\usage{
\method{coef}{stapDP}(object, ...)

\method{confint}{stapDP}(object, ...)

\method{fitted}{stapDP}(object, ...)

\method{nobs}{stapDP}(object, ...)

\method{VarCorr}{stapDP}(x)

\method{ngrps}{stapDP}(object, ...)

\method{nsamples}{stapDP}(object, ...)
}
\arguments{
\item{object}{stapDP object}

\item{...}{Ignored}

\item{x}{stapDP object}
}
\description{
Methods for stapDP objects
}
