% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stapDP_fit}
\alias{stapDP_fit}
\title{Functional Dirichlet Process Linear Regression}
\usage{
stapDP_fit(y, Z, X, w, tau_0, alpha_a, alpha_b, K, iter_max, burn_in, thin,
  seed, num_posterior_samples)
}
\arguments{
\item{y}{a vector of continuous outcomes}

\item{Z}{a matrix of population level confounders}

\item{X}{a matrix of spatial temporal aggregated predictors}

\item{w}{vector of weights for weighted regression}

\item{tau_0}{prior variance for STAP parameters}

\item{alpha_a}{alpha gamma prior hyperparameter}

\item{alpha_b}{alpha gamma prior hyperparameter}

\item{K}{truncation number}

\item{iter_max}{maximum number of iterations}

\item{burn_in}{number of burn in iterations}

\item{thin}{number by which to thin samples}

\item{seed}{rng initializer}

\item{num_posterior_samples}{number of final samples}
}
\description{
Functional Dirichlet Process Linear Regression
}
