% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapDPspec-methods.R
\name{stapDPspec-methods}
\alias{stapDPspec-methods}
\alias{get_k}
\alias{get_terms_ix}
\alias{get_component}
\alias{has_bw}
\alias{get_smooth_obj}
\title{\code{stapDPspec} Methods}
\usage{
get_k(x, term = NULL)

get_terms_ix(x, term = NULL)

get_component(x, term = NULL)

has_bw(x, term = NULL)

get_smooth_obj(x, term = NULL)
}
\arguments{
\item{x}{stapDPspec object}
}
\description{
\code{stapDPspec} Methods
}
\details{
Methods for stapDPspec objects. Function names are typically self-explanatory.
}
