% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapDPspec.R
\name{stapDPspec}
\alias{stapDPspec}
\title{Create STAP-DP data structure}
\usage{
stapDPspec(stapless_formula, fake_formula, stap_mat, K, benvo)
}
\arguments{
\item{stapless_formula}{from \code{\link{get_stapDPspec}}}

\item{fake_formula}{list of ``fake'' formulas from \code{\link{get_stapDPspec}}}

\item{stap_mat}{matrix of stap specification properties}

\item{K}{DP truncation}

\item{benvo}{Built Environment object - \code{\link[rbenvo]{benvo}} - containing data for model}
}
\description{
Create STAP-DP data structure
}
