% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stappDP_fit}
\alias{stappDP_fit}
\title{Penalized Functional Dirichlet Process Logistic Regression}
\usage{
stappDP_fit(y, Z, X, nu_0, alpha_a, alpha_b, K, iter_max, burn_in, thin,
  seed, num_posterior_samples)
}
\arguments{
\item{y}{vector of outcomes}

\item{Z}{matrix of subject covariates}

\item{X}{matrix of distance spline basis}
}
\description{
Penalized Functional Dirichlet Process Logistic Regression
}
