% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stappDP_mer_fit}
\alias{stappDP_mer_fit}
\title{Penalized Functional Dirichlet Process Linear Mixed Effects Regression}
\usage{
stappDP_mer_fit(
  y,
  Z,
  X,
  W,
  S,
  w,
  subj_mat_,
  subj_n,
  alpha_a,
  alpha_b,
  sigma_a,
  sigma_b,
  tau_a,
  tau_b,
  K,
  num_penalties,
  iter_max,
  burn_in,
  thin,
  seed,
  num_posterior_samples,
  fix_alpha
)
}
\arguments{
\item{y}{a vector of continuous outcomes}

\item{Z}{a matrix of population level confounders}

\item{X}{a matrix of spatial temporal aggregated predictors}

\item{W}{a design matrix for group specific terms}

\item{S}{penalty matrix for stap parameters}

\item{w}{a vector of weights for weighted regression}

\item{subj_mat_}{N x n sparse matrix used to aggregate subject observations}

\item{subj_n}{n x 1 vector of integers representing how many observations correspond to each subject}

\item{alpha_a}{alpha gamma prior shape hyperparameter}

\item{alpha_b}{alpha gamma prior scale hyperparameter}

\item{sigma_a}{precision gamma prior shape hyperparameter}

\item{sigma_b}{precision gamma prior scale hyperparameter}

\item{tau_a}{penalty gamma prior shape hyperparameter}

\item{tau_b}{penalty gamma prior scale hyperparameter}

\item{K}{truncation number}

\item{num_penalties}{number of penalty matrices accounted for in S}

\item{iter_max}{maximum number of iterations}

\item{burn_in}{number of burn in iterations}

\item{thin}{number by which to thin samples}

\item{seed}{rng initializer}

\item{num_posterior_samples}{total number of posterior samples}

\item{fix_alpha}{boolean value that determines whether or not to fix alpha in sampler}
}
\description{
fits a functional dirichlet process linear mixed effects regression model
with N observations and n subjects
}
