% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-summary.R
\name{summary.stapDP}
\alias{summary.stapDP}
\alias{print.summary.stapDP}
\title{Summary method for stapDP objects}
\usage{
\method{summary}{stapDP}(object, ..., digits = 1)

\method{print}{summary.stapDP}(x, digits = max(1, attr(x, "print.digits")), ...)
}
\arguments{
\item{object}{stapDP object}

\item{...}{ignored}

\item{digits}{Number of digits to use for formatting numbers when printing. 
When calling \code{summary}, the value of digits is stored as the 
\code{"print.digits"} attribute of the returned object.}

\item{x}{An object of class \code{"summary.stapDP"}.}
}
\value{
The \code{summary} method returns an object of class 
  \code{"summary.staDP"}  which is a matrix of 
  summary statistics and diagnostics, with attributes storing information for use by the
  \code{print} method.
}
\description{
Summaries of parameter estimates and MCMC convergence diagnostics 
(effective sample size, Rhat).
}
