% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.stapDP}
\alias{update.stapDP}
\title{Update STAPDP Model fit}
\usage{
\method{update}{stapDP}(
  object,
  formula,
  newdata,
  weights = NULL,
  iter_max = 1000,
  burn_in = floor(iter_max/2),
  chains = 1,
  alpha_a = 1,
  alpha_b = 1,
  sigma_a = 1,
  sigma_b = 1,
  tau_a = 1,
  tau_b = 1,
  K = 5,
  thin = 1,
  fix_alpha = FALSE,
  seed = NULL,
  scale = TRUE,
  center = TRUE,
  ...
)
}
\arguments{
\item{object}{stapDP object}

\item{formula}{formula used to specific fixed effects, see e.g. \code{\link[stats]{lm}}}

\item{newdata}{optional subject dataframe for subject level effects}
}
\description{
To save time from creating the data structures that are used in stapDP
models, one can "re-use" the previous data structure but change aspects of the 
fixed regression coefficients, hyperparameter or sampler settings.
}
