% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-summary.R
\name{diagnostics}
\alias{diagnostics}
\alias{diagnostics.stapDP}
\title{Diagnostics}
\usage{
diagnostics(x)

\method{diagnostics}{stapDP}(x)
}
\arguments{
\item{x}{a stapDP object}
}
\description{
Diagnostics
}
\section{Methods (by class)}{
\itemize{
\item \code{stapDP}: 
}}

