% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_staplm.R
\name{fdp_staplm.fit}
\alias{fdp_staplm.fit}
\title{Functional Dirichlet Process Spatial Temporal Aggregated Predictor Linear Model Fit}
\usage{
fdp_staplm.fit(
  y,
  Z,
  X,
  weights = rep(1, length(y)),
  alpha_a = 1,
  alpha_b = 1,
  sigma_a = 1,
  sigma_b = 1,
  tau_a = 1,
  tau_b = 1,
  K = 5L,
  iter_max,
  burn_in,
  thin = 1L,
  fix_alpha = FALSE,
  seed = NULL,
  chain = 1L,
  logging = FALSE,
  threshold = 0
)
}
\arguments{
\item{y}{vector of outcomes}

\item{Z}{design matrix}

\item{X}{stap design matrix}

\item{weights}{weights for weighted regression - default is vector of ones}

\item{alpha_a}{alpha gamma prior hyperparameter}

\item{alpha_b}{alpha gamma prior hyperparameter}

\item{sigma_a}{precision gamma prior hyperparameter}

\item{sigma_b}{precision gamma prior hyperparameter}

\item{tau_a}{penalty parameter gamma prior hyperparameter}

\item{tau_b}{penalty parameter gamma prior hyperparameter}

\item{K}{truncation number for DP mixture components}

\item{iter_max}{maximum number of iterations}

\item{burn_in}{number of iterations to burn-in}

\item{thin}{number by which to thin samples}

\item{fix_alpha}{boolean value}

\item{seed}{random number generator seed will be set to default value if not by user}

\item{chain}{chain label}

\item{logging}{boolean parameter indicating whether or not a single iteration should be run with print messages indicating successful completion of the Sampler's sub modules}

\item{threshold}{number of observations assigned to a cluster that may be considered negligble. Default is 0, but may need to be increased to 3,4, or 5 depending on how informative the data is.}
}
\description{
Functional Dirichlet Process Spatial Temporal Aggregated Predictor Linear Model Fit
}
