% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_staplm.R
\name{fdp_staplm}
\alias{fdp_staplm}
\title{Functional Dirichlet Process Spatial Temporal Aggregated Predictor in a Linear Model}
\usage{
fdp_staplm(y, Z, X, tau_0, alpha_a, alpha_b, K, iter_max, burn_in, thin,
  seed)
}
\arguments{
\item{y}{vector of outcomes}

\item{Z}{design matrix}

\item{X}{stap design matrix}

\item{tau_0}{normal base measure scale}

\item{alpha_a}{alpha gamma prior hyperparameter}

\item{alpha_b}{alpha gamma prior hyperparameter}

\item{K}{truncation number}

\item{burn_in}{number of burn in iterations}

\item{thin}{number by which to thin samples}

\item{seed}{rng initializer}

\item{iter_ix}{maximum number of iterations}
}
\description{
Functional Dirichlet Process Spatial Temporal Aggregated Predictor in a Linear Model
}
