% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_staplm.R
\name{fdp_staplm}
\alias{fdp_staplm}
\title{Functional Dirichlet Process Spatial Temporal Aggregated Predictor in a Linear Model}
\usage{
fdp_staplm(formula, subject_data, stap_term, dt_data = NULL, w = rep(1,
  nrow(subject_data)), tau_df = 1, alpha_a, alpha_b, K, penalize = T,
  iter_max, burn_in, thin = 1, knots = NULL, id_colname = "id",
  seed = NULL)
}
\arguments{
\item{formula}{Similar as for \code{\link[stats]{lm}}.}

\item{subject_data}{data frame of subject level measurements}

\item{stap_term}{call to \code{s_stap} with term in dt_data}

\item{dt_data}{data frame containing distances and/or times associated with subject data}

\item{w}{weights for weighted regression - default is vector of ones}

\item{tau_df}{if penalize = F: normal base measure scale 
if penalize true this is the degrees of freedom for the inverse chi square base measure.}

\item{alpha_a}{alpha gamma prior hyperparameter}

\item{alpha_b}{alpha gamma prior hyperparameter}

\item{K}{truncation number}

\item{penalize}{boolean that denotes whether parameters are penalized or not}

\item{iter_max}{maximum number of iterations}

\item{burn_in}{number of burn in iterations}

\item{thin}{number by which to thin samples}

\item{knots}{optional nots argument passed to \code{\link[mgcv]{smooth.construct}}.}

\item{id_colname}{string of id column name that is the unique id in both subject_data and dt_data}

\item{seed}{random number generator seed will be set to default value if not by user}
}
\description{
Functional Dirichlet Process Spatial Temporal Aggregated Predictor in a Linear Model
}
