% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_staplm.R
\name{fdp_staplm}
\alias{fdp_staplm}
\title{Functional Dirichlet Process Spatial Temporal Aggregated Predictor in a Linear Model}
\usage{
fdp_staplm(formula, knots, subject_data, dt_data = NULL, w = rep(1,
  nrow(subject_data)), tau_df = 1, alpha_a, alpha_b, K, penalize = T,
  iter_max, burn_in, thin = 1, seed = NULL)
}
\arguments{
\item{formula}{Similar as for \code{\link[stats]{lm}}.}

\item{knots}{quantiles of distance}

\item{alpha_a}{alpha gamma prior hyperparameter}

\item{alpha_b}{alpha gamma prior hyperparameter}

\item{K}{truncation number}

\item{penalize}{boolean that denotes whether parameters are penalized or not}

\item{iter_max}{maximum number of iterations}

\item{burn_in}{number of burn in iterations}

\item{thin}{number by which to thin samples}

\item{seed}{rng initializer}

\item{data}{frame}

\item{tau_0}{normal base measure scale}
}
\description{
Functional Dirichlet Process Spatial Temporal Aggregated Predictor in a Linear Model
}
