% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_staplmer.R
\name{fdp_staplmer.fit}
\alias{fdp_staplmer.fit}
\title{Functional Dirichlet Process Spatial Temporal Aggregated Predictor Linear Mixed Effects Regression Model Fit}
\usage{
fdp_staplmer.fit(
  y,
  Z,
  X,
  W,
  S,
  subj_mat,
  subj_n,
  weights = rep(1, length(y)),
  alpha_a = 1,
  alpha_b = 1,
  sigma_a = 1,
  sigma_b = 1,
  tau_a = 1,
  tau_b = 1,
  K = 5,
  iter_max,
  burn_in,
  thin = 1,
  fix_alpha = FALSE,
  seed = NULL
)
}
\arguments{
\item{y}{vector of outcomes}

\item{Z}{design matrix}

\item{X}{stap design matrix}

\item{W}{group terms design matrix from \code{\link[lme4]{glFormula}}}

\item{S}{list of penalty matrices from \code{\link[mgcv]{jagam}}}

\item{subj_mat}{matrix indexing subject-measurement locations in (Z,X,W)}

\item{subj_n}{vector of number of subject measurements}

\item{weights}{weights for weighted regression - default is vector of ones}

\item{alpha_a}{alpha gamma prior hyperparameter}

\item{alpha_b}{alpha gamma prior hyperparameter}

\item{sigma_a}{precision gamma prior hyperparameter}

\item{sigma_b}{precision gamma prior hyperparameter}

\item{tau_a}{penalty parameters gamma prior hyperparameter}

\item{tau_b}{penalty parameters gamma prior hyperparameter}

\item{K}{truncation number for DP mixture components}

\item{iter_max}{maximum number of iterations}

\item{burn_in}{number of iterations to burn-in}

\item{thin}{number by which to thin samples}

\item{fix_alpha}{boolean value}

\item{seed}{random number generator seed will be set to default value if not by user}
}
\description{
Functional Dirichlet Process Spatial Temporal Aggregated Predictor Linear Mixed Effects Regression Model Fit
}
