% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{FFR_benvo}
\alias{FFR_benvo}
\title{Samll benvo for use in \pkg{rstapDP} examples and vignettes.}
\format{
A \code{\link[rbenvo]{benvo}} consisting of the \code{FFR_subjects} and \code{FFR_distances} dataframes
}
\usage{
FFR_benvo
}
\description{
Samll benvo for use in \pkg{rstapDP} examples and vignettes.
}
\details{
see \code{\link{FFR_subjects}} and \code{\link{FFR_distances}}
For information on the dataframes.
}
\seealso{
\code{\link[rbenvo]{Benvo}} for information on benvos.
}
\keyword{datasets}
