% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{FFR_benvo}
\alias{FFR_benvo}
\title{Samll benvo for use in \pkg{rstapDP} examples and vignettes.}
\format{
A \code{\link[rbenvo]{benvo}} consisting of the \code{FFR_subjects} and \code{FFR_distances} dataframes

subject data: A data frame with 1000 rows and 3 columns
\describe{
\item{\code{id}}{The subject unique identifier}
\item{\code{sex}}{The measurement unique identifier}
\item{\code{BMI}}{The Built Environment Unique identifier}
}

Distance data: A data frame with 9501 rows and 2 columns
\describe{
\item{ \code{id}}{ The subject unique identifier}
\item{\code{Distance}}{The simulated distance between a hypothetical subject and fast food restaurant.}
}
}
\usage{
FFR_benvo
}
\description{
Samll benvo for use in \pkg{rstapDP} examples and vignettes.
}
\details{
see \code{\link{FFR_subjects}} and \code{\link{FFR_distances}}
For information on the dataframes.
}
\seealso{
\code{\link[rbenvo]{benvo}} for information on benvos.
}
\keyword{datasets}
