% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cluster_effects}
\alias{plot_cluster_effects}
\alias{plot_cluster_effects.stapDP}
\title{Cluster Spatial Temporal Effects}
\usage{
plot_cluster_effects(x, p = 0.95, switch = "color",
  prob_filter = 0.1, mode = T)

\method{plot_cluster_effects}{stapDP}(x, p = 0.95, switch = "color",
  prob_filter = 0.1, mode = T)
}
\arguments{
\item{x}{stapDP object}

\item{p}{probability contained in credible interval}

\item{switch}{one of "color" or "facet" for different plotting options}

\item{prob_filter}{all mixture components with median probability < prob_filter are excluded from the plot}

\item{mode}{boolean value indicating whether the cluster parameters that minimize the RSS should be used as a point estimate, or samples should be used.}
}
\value{
plot with cluster effect across space
}
\description{
Cluster Spatial Temporal Effects
}
\section{Methods (by class)}{
\itemize{
\item \code{stapDP}: 
}}

\keyword{internal}
