% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstapDP-package.R
\name{rstapDP-package}
\alias{rstapDP-package}
\alias{rstapDP}
\title{The `rstapDP package.`}
\description{
rstapDP is a package that implements clustered spatial-temporal aggregated predictor functions with the Dirichlet Process in R.
This allows for the modeling of features impact on measured subjects that can be related either through space or time as a function of some unobserved latent variable.
}
