% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapDP-methods.R
\name{stapDP-methods}
\alias{stapDP-methods}
\alias{VarCorr}
\alias{ngrps}
\alias{sigma}
\alias{nsamples}
\alias{coef.stapDP}
\alias{confint.stapDP}
\alias{fitted.stapDP}
\alias{nobs.stapDP}
\alias{VarCorr.stapDP}
\alias{ngrps.stapDP}
\alias{nsamples.stapDP}
\alias{ranef.stapDP}
\title{Methods for stapDP objects}
\usage{
\method{coef}{stapDP}(object, ...)

\method{confint}{stapDP}(object, ...)

\method{fitted}{stapDP}(object, ...)

\method{nobs}{stapDP}(object, ...)

\method{VarCorr}{stapDP}(x)

\method{ngrps}{stapDP}(object, ...)

\method{nsamples}{stapDP}(object, ...)

\method{ranef}{stapDP}(object, ...)
}
\arguments{
\item{object, }{x A fitted model object returned by one of the 
\pkg{rstapDP} modeling functions.}

\item{...}{Currently ignored.}

\item{x}{stapDP object}
}
\description{
Methods for stapDP objects that are similar to
their counterparts in the \pkg{stats} or \pkg{lme4} packages.
}
\details{
Methods for stapDP objects

Almost all methods behave intuitively as one familiar with the \pkg{stats} or \pkg{lme4} 
packages would expect.
}
