% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapDP.R
\name{stapDP}
\alias{stapDP}
\title{Create a stapDP object}
\usage{
stapDP(object, X_ranges)
}
\arguments{
\item{object}{A list provided by the nd_nhpp function}

\item{chains}{number of chains}

\item{J}{number of groups}
}
\value{
A stapDP object
}
\description{
Create a stapDP object
}
