% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapDP.R
\name{stapDP}
\alias{stapDP}
\title{Create a stapDP object}
\usage{
stapDP(object, X_ranges)
}
\arguments{
\item{object}{A list provided by the fdp_stap(p)lm function}

\item{X_ranges}{the range of the covariate space for the STAP}
}
\value{
A stapDP object
}
\description{
Create a stapDP object
}
