% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapDP.R
\name{stapDP}
\alias{stapDP}
\title{Create a stapDP object}
\usage{
stapDP(object)
}
\arguments{
\item{object}{A list provided by the fdp_staplm.fit function}
}
\value{
A stapDP object
}
\description{
Create a stapDP object
}
