% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stappDP_fit}
\alias{stappDP_fit}
\title{Penalized Functional Dirichlet Process Logistic Regression}
\usage{
stappDP_fit(y, Z, X, S, w, nu_0, alpha_a, alpha_b, K, iter_max, burn_in,
  thin, seed, num_posterior_samples)
}
\arguments{
\item{y}{a vector of continuous outcomes}

\item{Z}{a matrix of population level confounders}

\item{X}{a matrix of spatial temporal aggregated predictors}

\item{S}{penalty matrix for stap parameters}

\item{w}{a vector of weights for weighted regression}

\item{nu_0}{prior degrees of freedom for STAP regression coefficient scales}

\item{alpha_a}{alpha gamma prior hyperparameter}

\item{alpha_b}{alpha gamma prior hyperparameter}

\item{K}{truncation number}

\item{iter_max}{maximum number of iterations}

\item{burn_in}{number of burn in iterations}

\item{thin}{number by which to thin samples}

\item{seed}{rng initializer}

\item{num_posterior_samples}{total number of posterior samples}
}
\description{
Penalized Functional Dirichlet Process Logistic Regression
}
