% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stappDP_fit}
\alias{stappDP_fit}
\title{Penalized Functional Dirichlet Process Linear Regression with N observations}
\usage{
stappDP_fit(
  y,
  Z,
  X,
  w,
  alpha_a,
  alpha_b,
  sigma_a,
  sigma_b,
  tau_a,
  tau_b,
  K,
  threshold,
  iter_max,
  burn_in,
  thin,
  seed,
  num_posterior_samples,
  chain,
  fix_alpha,
  logging
)
}
\arguments{
\item{y}{a vector of continuous outcomes}

\item{Z}{a matrix of population level confounders}

\item{X}{a matrix of spatial temporal aggregated predictors}

\item{w}{a vector of weights for weighted regression}

\item{alpha_a}{alpha gamma prior shape hyperparameter}

\item{alpha_b}{alpha gamma prior scale hyperparameter}

\item{sigma_a}{precision gamma prior shape hyperparameter}

\item{sigma_b}{precision gamma prior scale hyperparameter}

\item{tau_a}{penalty gamma prior shape hyperparameter}

\item{tau_b}{penalty gamma prior scale hyperparameter}

\item{K}{truncation number}

\item{threshold}{number of members per cluster at which cluster is included in regression}

\item{iter_max}{maximum number of iterations}

\item{burn_in}{number of burn in iterations}

\item{thin}{number by which to thin samples}

\item{seed}{rng initializer}

\item{num_posterior_samples}{total number of posterior samples}

\item{chain}{chain label}

\item{fix_alpha}{boolean value that determines whether or not to fix alpha in sampler}

\item{logging}{boolean parameter indicating whether or not a single iteration should be run with print messages indicating successful completion of the Sampler's sub modules}
}
\description{
Penalized Functional Dirichlet Process Linear Regression with N observations
}
