% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{traceplots}
\alias{traceplots}
\alias{traceplots.stapDP}
\title{Diagnostic Traceplots}
\usage{
traceplots(x, par = NULL)

\method{traceplots}{stapDP}(x, par = NULL)
}
\arguments{
\item{x}{a stapDP object}

\item{par}{string of parameter to use
default is .1}
}
\description{
Diagnostic Traceplots
}
\section{Methods (by class)}{
\itemize{
\item \code{stapDP}: 
}}

